Refactpred goTo
