+++ src/lib/tests.js
@@ -6,6 +6,7 @@ const { setEnvs, log, timestamp } = require("./utils");
 const util = require("util");
 const exec = util.promisify(require("child_process").exec);
 const axios = require("axios");
+const { goTo } = require("./tests/goTo");
 const { moveMouse } = require("./tests/moveMouse");
 const { scroll } = require("./tests/scroll");
 const { screenshot } = require("./tests/screenshot");
@@ -205,7 +206,7 @@ async function runAction(config, action, page, videoDetails) {
   result.result = {};
   switch (action.action) {
     case "goTo":
-      result = await openUri(action, page);
+      result = await goTo(action, page);
       break;
     case "find":
       // Perform sub-action: wait
@@ -610,46 +611,4 @@ async function findElement(action, page) {
     let result = { status, description };
     return { result, elementHandle };
   }
-}
-
-// Open a URI in the browser
-async function openUri(action, page) {
-  let uri;
-  if (!action.uri) {
-    // FAIL: No URI
-    let status = "FAIL";
-    let description = "'uri' is a required field.";
-    let result = { status, description };
-    return { result };
-  }
-  // Load environment variables
-  if (action.env) {
-    let result = await setEnvs(action.env);
-    if (result.status === "FAIL") return { result };
-  }
-  if (
-    action.uri[0] === "$" &&
-    process.env[action.uri.substring(1)] != undefined
-  ) {
-    uri = process.env[action.uri.substring(1)];
-  } else {
-    uri = action.uri;
-  }
-  // Catch common formatting errors
-  if (!uri.includes("://")) uri = "https://" + uri;
-  // Run action
-  try {
-    await page.goto(uri);
-  } catch {
-    // FAIL: Error opening URI
-    let status = "FAIL";
-    let description = "Couldn't open URI.";
-    let result = { status, description };
-    return { result };
-  }
-  // PASS
-  let status = "PASS";
-  let description = "Opened URI.";
-  let result = { status, description };
-  return { result };
 }
